[Configs] Fix config env overrides
Given that `list_separator` is called, the `config` crate expects every key to be a list unless `with_list_parse_key` is called which defines the list of keys that are supposed to be vectors. This diff lets each service define which of its keys are vector and pass it during service config registration.